# vegro-hr/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payroll API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "List companies",
                "description": "Get all companies (super admin only)",
                "operationId": "listCompanies",
                "responses": {
                    "200": {
                        "description": "Companies retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Companies retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create company",
                "description": "Create a company and optional demo data (super admin only)",
                "operationId": "createCompany",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "admin_name",
                                    "admin_email",
                                    "admin_password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Ltd"
                                    },
                                    "domain": {
                                        "type": "string",
                                        "example": "acme.local",
                                        "nullable": true
                                    },
                                    "plan": {
                                        "type": "string",
                                        "example": "starter"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "environment": {
                                        "type": "string",
                                        "example": "demo"
                                    },
                                    "seed_demo": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "admin_name": {
                                        "type": "string",
                                        "example": "Company Admin"
                                    },
                                    "admin_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@acme.local"
                                    },
                                    "admin_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/{company}": {
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update company",
                "description": "Update company profile fields (super admin only)",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acme Ltd"
                                    },
                                    "industry": {
                                        "type": "string",
                                        "example": "Manufacturing"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "KE"
                                    },
                                    "environment": {
                                        "type": "string",
                                        "example": "production"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}